pyproject.toml
[tool.poetry]
name = "avulsionprecursors"
version = "0.1.0"
description = "Tools for analyzing river avulsion precursors using BASED and ICESat-2 data"
authors = ["Jake Gearon <jake.gearon@gmail.com>"]
readme = "README.md"
packages = [{include = "avulsionprecursors"}]

[tool.poetry.dependencies]
python = "^3.10"
pandas = "^2.0.0"
geopandas = "^0.14.0"
xgboost = "^2.0.0"
earthengine-api = "^0.1.0"
pygeodesy = "^23.0.0"
matplotlib = "^3.0.0"
scipy = "^1.0.0"
statsmodels = "^0.14.0"

python-dotenv = "^1.0.1"
sqlalchemy = "^2.0.36"
pyqt5 = "^5.15.11"
cartopy = "^0.24.1"
psycopg2 = "2.9.10"
sliderule = "^4.9.4"
[tool.poetry.group.dev.dependencies]
pytest = "^7.0.0"
black = "^23.0.0"
isort = "^5.0.0"
mypy = "^1.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
avulsion-process = "avulsionprecursors.cli:main"
